+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2019-07-15
+### Added
+- `API base` - getInstance helper for propagate client
+
+### Removed
+- `Controller` functionality
+
 ## [1.7.0] - 2019-07-11
 ### Added
 - ENV `MS_PATH` - path prefix